+++ src/game/GameObjects/BetonWall.h
@@ -39,7 +39,7 @@ public:
 
     BetonWall(const EBetonWallType eBetonWallType, const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
 	virtual void Render() const override;
-	virtual void Update(const uint64_t delta) override;
+	virtual void Update(const double delta) override;
 
 private:
     std::array<EBlockState, 4> eCurrentBlockState_;
