+++ src/game/GameObjects/Water.h
@@ -23,7 +23,7 @@ public:
 
     Water(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
     virtual void Render() const override;
-    void Update(const uint64_t delta) override;
+    void Update(const double delta) override;
 
 private:
     std::shared_ptr<RenderEngine::Sprite> sprite_;
