+++ src/renderer/Sprite.cpp
@@ -111,7 +111,7 @@ namespace RenderEngine {
 		Renderer::Draw(vertexArray_, indexBuffer_, *pShaderProgram_);
 	}
 
-	uint64_t Sprite::GetFrameDuration(const size_t frameId) const
+	double Sprite::GetFrameDuration(const size_t frameId) const
 	{
 		return framesDescriptions_[frameId].duration;
 	}
