+++ src/renderer/Sprite.h
@@ -21,7 +21,7 @@ namespace RenderEngine {
 
 		struct FrameDescription {
 			
-			FrameDescription(const glm::vec2 _leftBottomUV, const glm::vec2 _rightTopUV, const uint64_t _duration)
+			FrameDescription(const glm::vec2 _leftBottomUV, const glm::vec2 _rightTopUV, const double _duration)
 				: leftBottomUV(_leftBottomUV)
 				, rightTopUV(_rightTopUV)
 				, duration(_duration)
@@ -29,7 +29,7 @@ namespace RenderEngine {
 			
 			glm::vec2 leftBottomUV;
 			glm::vec2 rightTopUV;
-			uint64_t duration;
+			double duration;
 		};
 
 		Sprite(std::shared_ptr<Texture2D> pTexture,
@@ -43,7 +43,7 @@ namespace RenderEngine {
 
 		void Render(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer = 0.f, const size_t frameId = 0) const;
 
-		uint64_t GetFrameDuration(const size_t frameId) const;
+		double GetFrameDuration(const size_t frameId) const;
 		size_t GetFramesCount() const;
 		void InsertFrames(std::vector<FrameDescription> framesDescriptions);
