+++ src/renderer/SpriteAnimator.h
@@ -12,13 +12,13 @@ namespace RenderEngine {
 		SpriteAnimator(std::shared_ptr<Sprite> pSprite);
 
 		size_t GetCurrentFrame() const;
-		void Update(const uint64_t delta);
+		void Update(const double delta);
 
 	private:
 		std::shared_ptr<Sprite> pSprite_;
 		size_t current_frame_;
-		uint64_t current_frame_duration_;
-		uint64_t current_animation_time_;
+		double current_frame_duration_;
+		double current_animation_time_;
 	};
 
 }
